youtube_Notes/requirements.txt
aioitertools==0.11.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==41.0.4
decorator==5.1.1
distlib==0.3.7
distro==1.9.0
dulwich==0.21.6
executing==2.0.1
fastapi==0.109.2
fastjsonschema==2.18.1
filelock==3.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
ipython==8.22.1
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
keyring==24.2.0
matplotlib-inline==0.1.6
more-itertools==10.1.0
msgpack==1.0.7
openai==1.12.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pipenv==2023.11.17
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pyproject_hooks==1.0.0
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.36.3
tiktoken==0.6.0
tomlkit==0.12.1
tqdm==4.66.2
traitlets==5.14.1
trove-classifiers==2023.9.19
typing_extensions==4.9.0
urllib3==2.0.6
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
youtube-transcript-api==0.6.2
zipp==3.17.0
